+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Begin adding support for multiple inputs and outputs
 * Added `MediaConversion#getFormats(callback)` which calls back with available formats & codecs
 * Added `getRawFramesOutput(pixel_format)` so you can just decode a stream
+* Added `setMany(options)` to `Input` and `Output` objects
 
 ## 0.1.0 (2016-10-19)